# api/openapi/categories.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Our Expenses server
  description: TODO
servers:
  - url: https://{hostname}/api
    variables:
      hostname:
        default: localhost
security:
  # - bearerAuth: []
paths:
  /categories:
    get:
      summary: Returns all categories
      description: |
        Returns all categories from the system.
      operationId: findCategories
      parameters:
        - name: parentId
          in: query
          description: category parentId to filter by
          required: false
          schema:
            type: string
        - name: allChildren
          in: query
          description: include all category children
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: categories response
          content:
            application/json:
              schema:
                type: array
                $ref: '#/components/schemas/Category'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Creates a new category
      description: Creates a new category in the system.
      operationId: addCategory
      requestBody:
        description: Category to add to the system
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategory'
      responses:
        '200':
          description: Category response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /categories/{id}:
    get:
      summary: Returns a category by ID
      description: Returns a category based on a single ID
      operationId: findCategoryByID
      parameters:
        - name: id
          in: path
          description: ID of category to fetch
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: category response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Updates category
      description: Updates category in the system.
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          description: ID of category to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Category to update in the system
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Category:
      allOf:
        - $ref: '#/components/schemas/NewCategory'
        - required:
          - id
          properties:
            id:
              type: string
              format: uuid
              description: Unique id of the category
    NewCategory:
      type: object
      required:
        - name
        - path
        - level
      properties:
        name:
          type: string
          description: Name of the category
        parentId:
          type: string
          format: uuid
        path:
          type: string
        level:
          type: integer
        parents:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message