# internal/categories/ports/openapi_api.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package ports

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/labstack/echo/v4"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Returns all categories
	// (GET /categories)
	FindCategories(ctx echo.Context, params FindCategoriesParams) error
	// Creates a new category
	// (POST /categories)
	AddCategory(ctx echo.Context) error
	// Returns a category by ID
	// (GET /categories/{id})
	FindCategoryByID(ctx echo.Context, id string) error
	// Updates category
	// (PUT /categories/{id})
	UpdateCategory(ctx echo.Context, id string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// FindCategories converts echo context to params.
func (w *ServerInterfaceWrapper) FindCategories(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params FindCategoriesParams
	// ------------- Optional query parameter "parentId" -------------

	err = runtime.BindQueryParameter("form", true, false, "parentId", ctx.QueryParams(), &params.ParentId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter parentId: %s", err))
	}

	// ------------- Optional query parameter "allChildren" -------------

	err = runtime.BindQueryParameter("form", true, false, "allChildren", ctx.QueryParams(), &params.AllChildren)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter allChildren: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.FindCategories(ctx, params)
	return err
}

// AddCategory converts echo context to params.
func (w *ServerInterfaceWrapper) AddCategory(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.AddCategory(ctx)
	return err
}

// FindCategoryByID converts echo context to params.
func (w *ServerInterfaceWrapper) FindCategoryByID(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.FindCategoryByID(ctx, id)
	return err
}

// UpdateCategory converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateCategory(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.UpdateCategory(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/categories", wrapper.FindCategories)
	router.POST(baseURL+"/categories", wrapper.AddCategory)
	router.GET(baseURL+"/categories/:id", wrapper.FindCategoryByID)
	router.PUT(baseURL+"/categories/:id", wrapper.UpdateCategory)

}