# internal/categories/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package ports

// Category defines model for Category.
type Category struct {
	// Embedded struct due to allOf(#/components/schemas/NewCategory)
	NewCategory `yaml:",inline"`
	// Embedded fields due to inline allOf schema

	// Unique id of the category
	Id string `json:"id"`
}

// Error defines model for Error.
type Error struct {

	// Error code
	Code int32 `json:"code"`

	// Error message
	Message string `json:"message"`
}

// NewCategory defines model for NewCategory.
type NewCategory struct {
	Level int `json:"level"`

	// Name of the category
	Name     string      `json:"name"`
	ParentId *string     `json:"parentId,omitempty"`
	Parents  *[]Category `json:"parents,omitempty"`
	Path     string      `json:"path"`
}

// FindCategoriesParams defines parameters for FindCategories.
type FindCategoriesParams struct {

	// category parentId to filter by
	ParentId *string `json:"parentId,omitempty"`

	// include all category children
	AllChildren *bool `json:"allChildren,omitempty"`
}

// AddCategoryJSONBody defines parameters for AddCategory.
type AddCategoryJSONBody NewCategory

// UpdateCategoryJSONBody defines parameters for UpdateCategory.
type UpdateCategoryJSONBody Category

// AddCategoryJSONRequestBody defines body for AddCategory for application/json ContentType.
type AddCategoryJSONRequestBody AddCategoryJSONBody

// UpdateCategoryJSONRequestBody defines body for UpdateCategory for application/json ContentType.
type UpdateCategoryJSONRequestBody UpdateCategoryJSONBody